Committing this fix to main because, as my greatest idol minister Faber once said, "Ik ben beleid".
Small fix to the prices of packages when sending a package.

Price was not stored correctly, prices were rounded down in conversion from double to int.

Prices are not stored in ints and should be converted to doubles/strings. SendPackageController.dart has getPriceAsString() method doing this.

So price "12,95" is now stored in backend as: "1295".